Removing wintpy folder from src, and switching to monkey patching instead
