+++ boiiiwd_package/src/imports.py
@@ -32,7 +32,7 @@ from .CTkListbox.ctk_listbox import CTkListbox
 from .CTkToolTip.ctk_tooltip import CTkToolTip
 
 # winpty
-from .winpty import PtyProcess
+from src.winpty_patch import PtyProcess
 import ctypes
 
 if getattr(sys, 'frozen', False):
