--- boiiiwd_package/src/winpty/__init__.py
@@ -1,23 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Pywinpty
-========
-This package provides low and high level APIs to create
-pseudo terminals in Windows.
-"""
-
-# Local imports
-from .winpty import PTY, WinptyError, __version__
-from .ptyprocess import PtyProcess
-from .enums import Backend, Encoding, MouseMode, AgentConfig
-
-
-PTY
-PtyProcess
-Backend
-Encoding
-MouseMode
-AgentConfig
-WinptyError
-
-__version__
